Green Light |

+++ KZF0019/CA02/prod/StarSensor.py
@@ -65,8 +65,8 @@ class StarSensor(object):
             #print 'a'
             return None
         else:
-            rightAscension =  float(math.pi / 2 + self.env.getTime() * 2 * math.pi / (86164.1*1000000))
-            declination =  float(self.env.getTime() * 2 * math.pi / self.env.getRotationalPeriod())
+            rightAscension =  math.pi / 2 + self.env.getTime() * 2 * math.pi / (86164.1*1000000)
+            declination = self.env.getTime() * 2 * math.pi / self.env.getRotationalPeriod()
             
             while (rightAscension-self.fieldView/2) >= 2 * math.pi:
                 rightAscension = rightAscension - 2 * math.pi
